+++ recipes/packages.rb
@@ -1,3 +1,0 @@
-node['htpasswd']['packages'].each do |pck|
-  package pck
-end