+++ test/integration/delete/inspec/delete_spec.rb
@@ -1,0 +1,4 @@
+describe file '/tmp/htpass_test' do
+  its('content') { should_not match /^admin/ }
+  its('mode') { should cmp '0644' }
+end