+++ CHANGELOG.md
@@ -1,5 +1,73 @@
-# Changelog
+# htpasswd Cookbook Changelog
 
-## UNRELEASED
+This file is used to list changes made in each version of the htpasswd cookbook.
 
-- Adopted by Sous-Chefs
+## Unreleased
+
+### Fixed
+
+- Automated PR: Cookstyle Changes [\#39](https://github.com/sous-chefs/htpasswd/pull/39)
+- Automated PR: Standardising Files [\#38](https://github.com/sous-chefs/htpasswd/pull/38)
+- Fix converge failures on Chef 16
+- Fix idempotency for `:add` and `:delete` actions
+- Fix htauth gem loading issues
+
+### Added
+
+- Support for 'plaintext' type [\#32](https://github.com/sous-chefs/htpasswd/pull/32)
+- Adoption by Sous-Chefs [\#40](https://github.com/sous-chefs/htpasswd/pull/40)
+- Add delete and overwrite test-kitchen suites
+- Add mode parameters for file
+
+### Changed
+
+- Update test-kitchen to modern platforms
+- Update metadata.rb for Sous Chefs
+- Update README.md for Sous Chefs
+- Convert htpasswd to a custom resource
+
+### Removed
+
+- Remove unused recipes and other dependencies
+
+## 0.3.0 - 2017-05-21
+
+- Tests [\#30](https://github.com/sous-chefs/htpasswd/pull/30)
+- Fix FC069, FC071, FC078 [\#26](https://github.com/sous-chefs/htpasswd/pull/26)
+- Update testing [\#25](https://github.com/sous-chefs/htpasswd/pull/25)
+- Switch from python to poise-python [\#24](https://github.com/sous-chefs/htpasswd/pull/24)
+- Update attribute default.rb  [\#22](https://github.com/sous-chefs/htpasswd/pull/22)
+- Drop ruby 1.9.3 [\#20](https://github.com/sous-chefs/htpasswd/pull/20)
+- Add default package to packages attribute [\#19](https://github.com/sous-chefs/htpasswd/pull/19)
+- Fix Jenkins URL to match github repo [\#18](https://github.com/sous-chefs/htpasswd/pull/18)
+
+## 0.2.4 - 2014-09-23
+
+- Remove ruby dep (very old and unmaintained) (fixed #15)
+- Fix rubocop
+- Update rubocop
+
+## 0.2.2 - 2014-09-17
+
+- Cleanup provider init
+- Add why-run
+
+## 0.2.0 - 2014-09-17
+
+- Add matchers for chefspec [\#14](https://github.com/sous-chefs/htpasswd/pull/14)
+- Added a new resource type so the hashing algorithm can be changed [\#13](https://github.com/sous-chefs/htpasswd/pull/13)
+
+## 0.1.2 - 2014-03-14
+
+- Rubocop [\#12](https://github.com/sous-chefs/htpasswd/pull/12)
+- Fix travis [\#11](https://github.com/sous-chefs/htpasswd/pull/11)
+- Use htauth lib to manage htpasswd files. [\#10](https://github.com/sous-chefs/htpasswd/pull/10)
+- Added :delete action for user [\#5](https://github.com/sous-chefs/htpasswd/pull/5)
+
+## 0.0.3 - 2013-07-18
+
+- Berkshelf support
+
+## 0.0.2 - 2013-07-18
+
+- Initial release